+++ lib/heap/constants.js
