+++ lib/heap/helper.js
