+++ lib/heap/index.js
@@ -52,7 +52,7 @@ module.exports = {
     },
 
     isSupported: function() {
-        return $.when(true);
+        return true;
     },
 
     init: function(metas) {
