+++ lib/indexed-db/helper.js
