+++ lib/indexed-db/index.js
@@ -24,18 +24,6 @@ var cursorReader = function cursorReader(result) {
         _this.resolve(_this.data);
     }
 };
-//exclude IE till 10 and safari
-function shouldExcludeBrowser() {
-    var ua = window.navigator.userAgent.toLowerCase();
-    if (ua.indexOf('msie') > -1) {
-        return true;
-    } else if (ua.indexOf('safari') > -1 && ua.indexOf('chrome') === -1) {
-        return true;
-    }
-
-    return false;
-}
-
 //FIX me - chk for specified table availability
 module.exports = {
     db: null,
@@ -47,9 +35,7 @@ module.exports = {
 
     isSupported: function() {
         //check if we have indexedDB support or not
-        // FIXME: can use !! (double not)
-        var isSupported = (window.indexedDB || window.mozIndexedDB || window.webkitIndexedDB || window.msIndexedDB);
-        return $.when(isSupported && !shouldExcludeBrowser());
+        return !!(window.indexedDB || window.mozIndexedDB || window.webkitIndexedDB || window.msIndexedDB);
     },
 
     //create DB and initialize with object stores
@@ -230,8 +216,9 @@ module.exports = {
     //delete the created db
     destroy: function(id) {
         var deferred = new $.Deferred();
+        var deleteRequest;
         if (!this.id) {
-            var deleteRequest = indexedDB.deleteDatabase(id);
+            deleteRequest = indexedDB.deleteDatabase(id);
             deleteRequest.onerror = deferred.reject;
             deleteRequest.onsuccess = deferred.resolve;
             return deferred.promise();
@@ -239,7 +226,7 @@ module.exports = {
 
         this.db.close();
         this.db = null;
-        var deleteRequest = indexedDB.deleteDatabase(this.id);
+        deleteRequest = indexedDB.deleteDatabase(this.id);
         this.id = null;
         deleteRequest.onerror = deferred.reject;
         deleteRequest.onsuccess = deferred.resolve;
