+++ package.json
@@ -1,8 +1,8 @@
 {
     "name": "localdatastore",
     "version": "0.1.7",
-    "description": "Promise based client/server side storage. Stores any data. Index based querying supported. Uses IndexedDB/heap based on availability",
-    "main": "index.js",
+    "description": "Promise based client/server side storage. Uses IndexedDB/heap based on availability",
+    "main": "lib/index.js",
     "scripts": {
         "test": "grunt test"
     },
@@ -34,36 +34,30 @@
         }
     ],
     "dependencies": {
-        "jquery": "^1.8.3"
+        "jquery": "latest"
     },
     "devDependencies": {
-        "lasso": "^1",
-        "grunt": "^0.4",
-        "grunt-config-dir": "^0",
-        "grunt-karma": "^0.10.1",
-        "karma-chai": "^0.1.0",
-        "karma-mocha": "^0.1.10",
-        "karma-lasso": "^1",
-        "lasso-require": "^1",
-        "lasso-less": "^2",
-        "lasso-dust": "^1",
-        "karma-phantomjs-launcher": "^0.1.4",
-        "karma-mocha-reporter": "^0.3.1",
-        "karma-chrome-launcher": "^0.1.7",
-        "karma-firefox-launcher": "^0.1.4",
-        "karma-safari-launcher": "^0.1.1",
-        "karma-opera-launcher": "^0.1.0",
-        "karma-ie-launcher": "^0.1.5",
-        "karma-webdriver-launcher": "^1.0",
-        "glob": "^4.3.5",
-        "chai": "^1",
-        "chai-a11y": "^0",
-        "mocha": "^1",
-        "supertest": "^0",
-        "sinon": "1.14.1",
-        "sinon-chai": "^2.7.0",
-        "karma-sinon": "^1.0.4",
-        "karma-sinon-expect": "^0.1.4",
-        "karma-coverage": "^0.2"
+        "lasso": "latest",
+        "grunt": "latest",
+        "grunt-karma": "latest",
+        "karma-chai": "latest",
+        "karma-mocha": "latest",
+        "karma-lasso": "latest",
+        "lasso-require": "latest",
+        "karma-phantomjs-launcher": "latest",
+        "karma-mocha-reporter": "latest",
+        "karma-chrome-launcher": "latest",
+        "karma-firefox-launcher": "latest",
+        "karma-safari-launcher": "latest",
+        "karma-opera-launcher": "latest",
+        "karma-ie-launcher": "latest",
+        "karma-webdriver-launcher": "latest",
+        "grunt-contrib-jshint": "latest",
+        "grunt-contrib-clean": "latest",
+        "glob": "latest",
+        "chai": "latest",
+        "mocha": "latest",
+        "sinon": "latest",
+        "sinon-chai": "latest"
     }
 }
