+++ test/indexed-db/helper.js
@@ -1,20 +1,14 @@
 'use strict';
-require('../../../data-store/shims')
-var idb = require('../../../data-store/indexed-db');
-var helper = require('../../../data-store/indexed-db/helper');
+var idb = require('../../lib/indexed-db');
+var helper = require('../../lib/indexed-db/helper');
 var expect = require('chai').expect;
 var isSupported = false;
 describe('data-store/indexed-db/helper', function() {
-    before(function(done) {
-        return idb.isSupported().then(function(supported) {
-            isSupported = supported;
-            done();
-        });
-    });
+    if (!idb.isSupported()) {
+        return;
+    }
+
     it('creates db without any errors', function() {
-        if (!isSupported) {
-            return;
-        }
         return helper.createDB('db1', [
             {
                 indexes: [{name: 'attr1'}],
