+++ test/indexed-db/insert.js
@@ -1,19 +1,12 @@
 'use strict';
-require('../../../data-store/shims')
-var idb = require('../../../data-store/indexed-db');
+var idb = require('../../lib/indexed-db');
 var expect = require('chai').expect;
-var isSupported = false;
 describe('data-store/indexed-db/insert', function() {
-    before(function(done) {
-        return idb.isSupported().then(function(supported) {
-            isSupported = supported;
-            done();
-        });
-    });
+    if (!idb.isSupported()) {
+        return;
+    }
+
     it('lets you insert valid values in the database', function() {
-        if (!isSupported) {
-            return;
-        }
         return idb.init('db1-insert', [
             {
                 indexes: [{name: 'attr1'}],
@@ -38,9 +31,6 @@ describe('data-store/indexed-db/insert', function() {
         });
     });
     it('doesn\'t allow inserting in a non-existing table', function(done) {
-        if (!isSupported) {
-            done();
-        }
         return idb.init('db2-insert', [
             {
                 indexes: [{name: 'attr1'}],
@@ -60,9 +50,6 @@ describe('data-store/indexed-db/insert', function() {
         });
     });
     it('doesnt let you insert values with missing values for indexes', function() {
-        if (!isSupported) {
-            return;
-        }
         return idb.init('db3-insert', [
             {
                 indexes: [{name: 'attr1'}],
