+++ test/indexed-db/remove.js
@@ -1,16 +1,20 @@
 'use strict';
-require('../../../data-store/shims');
-var heap = require('../../../data-store/heap');
+var idb = require('../../lib/indexed-db');
 var expect = require('chai').expect;
-describe('data-store/heap/remove', function() {
+var isSupported = false;
+describe('data-store/indexed-db/remove', function() {
+    if (!idb.isSupported()) {
+        return;
+    }
+
     it('removes the value based on one key in an index', function() {
-        return heap.init([
+        return idb.init('db1-remove', [
             {
                 indexes: [{name: 'attr1'}],
                 name: 'table1'
             }
         ]).then(function() {
-            return heap.insert({name: 'table1'}, [
+            return idb.insert({name: 'table1'}, [
                 {
                     attr1: 'A',
                     attr2: 'B'
@@ -20,30 +24,31 @@ describe('data-store/heap/remove', function() {
                     attr2: 'B2'
                 }
             ]).then(function() {
-                return heap.remove({name: 'table1'}, {
+                return idb.remove({name: 'table1'}, {
                     index: 'attr1',
                     data: ['A']
                 }).then(function() {
-                    return heap.select({
+                    return idb.select({
                         index: 'attr1',
                         name: 'table1'
                     }, {}).then(function(data) {
                         expect(data.length).to.eql(1);
                         expect(data[0].attr1).to.eql('A2');
                         expect(data[0].attr2).to.eql('B2');
+                        return idb.destroy('db1-remove');
                     });
                 });
             });
         });
     });
     it('removes the value based on an array of keys in an index', function() {
-        return heap.init([
+        return idb.init('db2-remove', [
             {
                 indexes: [{name: 'attr1'}],
                 name: 'table1'
             }
         ]).then(function() {
-            return heap.insert({name: 'table1'}, [
+            return idb.insert({name: 'table1'}, [
                 {
                     attr1: 'A',
                     attr2: 'B'
@@ -53,31 +58,32 @@ describe('data-store/heap/remove', function() {
                     attr2: 'B2'
                 }
             ]).then(function() {
-                return heap.remove({name: 'table1'}, {
+                return idb.remove({name: 'table1'}, {
                     index: 'attr1',
                     data: [
                         'A',
                         'A2'
                     ]
                 }).then(function() {
-                    return heap.select({
+                    return idb.select({
                         index: 'attr1',
                         name: 'table1'
                     }, {}).then(function(data) {
                         expect(data.length).to.eql(0);
+                        return idb.destroy('db2-remove');
                     });
                 });
             });
         });
     });
     it('removes all data if there is no filter', function() {
-        return heap.init([
+        return idb.init('db3-remove', [
             {
                 indexes: [{name: 'attr1'}],
                 name: 'table1'
             }
         ]).then(function() {
-            return heap.insert({name: 'table1'}, [
+            return idb.insert({name: 'table1'}, [
                 {
                     attr1: 'A',
                     attr2: 'B'
@@ -85,17 +91,19 @@ describe('data-store/heap/remove', function() {
                 {
                     attr1: 'A2',
                     attr2: 'B2'
+                },
+                {
+                    attr1: 'A3',
+                    attr2: 'B3'
                 }
             ]).then(function() {
-                return heap.remove({name: 'table1'}, {}).then(function() {
-                    var metadata = {};
-                    metadata.index = 'attr1';
-                    metadata.name = 'table1';
-                    return heap.select({
+                return idb.remove({name: 'table1'}, {}).then(function() {
+                    return idb.select({
                         index: 'attr1',
                         name: 'table1'
                     }, {}).then(function(data) {
                         expect(data.length).to.eql(0);
+                        return idb.destroy('db3-remove');
                     });
                 });
             });
@@ -103,13 +111,13 @@ describe('data-store/heap/remove', function() {
     });
     it('throws an error if an remove is attempted on a non existing table', function(done) {
         //to be fixed
-        return heap.init([
+        return idb.init('db4-remove', [
             {
                 indexes: [{name: 'attr1'}],
                 name: 'table1'
             }
         ]).then(function() {
-            return heap.insert({name: 'table1'}, [
+            return idb.insert({name: 'table1'}, [
                 {
                     attr1: 'A',
                     attr2: 'B'
@@ -119,21 +127,22 @@ describe('data-store/heap/remove', function() {
                     attr2: 'B2'
                 }
             ]).then(function() {
-                return heap.remove({name: 'invalidtable'}, {}).fail(function(resp) {
-                    expect(resp).to.eql('Meta not found');
-                    done();
+                return idb.remove({name: 'invalidtable'}, {}).fail(function() {
+                    return idb.destroy('db1-remove').then(function() {
+                        done();
+                    });
                 });
             });
         });
     });
     it('doesn\'t remove anything if key search in index doesn\'t fetch any data', function() {
-        return heap.init([
+        return idb.init('db5-remove', [
             {
                 indexes: [{name: 'attr1'}],
                 name: 'table1'
             }
         ]).then(function() {
-            return heap.insert({name: 'table1'}, [
+            return idb.insert({name: 'table1'}, [
                 {
                     attr1: 'A',
                     attr2: 'B'
@@ -143,15 +152,16 @@ describe('data-store/heap/remove', function() {
                     attr2: 'B2'
                 }
             ]).then(function() {
-                return heap.remove({name: 'table1'}, {
+                return idb.remove({name: 'table1'}, {
                     index: 'attr1',
                     data: ['C']
                 }).then(function() {
-                    return heap.select({
+                    return idb.select({
                         index: 'attr1',
                         name: 'table1'
                     }, {}).then(function(data) {
                         expect(data.length).to.eql(2);
+                        return idb.destroy('db5-remove');
                     });
                 });
             });
