+++ test/indexed-db/update.js
@@ -1,19 +1,12 @@
 'use strict';
-require('../../../data-store/shims')
-var idb = require('../../../data-store/indexed-db');
+var idb = require('../../lib/indexed-db');
 var expect = require('chai').expect;
-var isSupported = false;
 describe('data-store/indexed-db/update', function() {
-    before(function(done) {
-        return idb.isSupported().then(function(supported) {
-            isSupported = supported;
-            done();
-        });
-    });
+    if (!idb.isSupported()) {
+        return;
+    }
+
     it('updates the value based on one key in an index', function() {
-        if (!isSupported) {
-            return;
-        }
         return idb.init('db1-update', [
             {
                 indexes: [{name: 'attr1'}],
@@ -58,9 +51,6 @@ describe('data-store/indexed-db/update', function() {
     });
 
     it('updates the value based on an array of keys in an index', function() {
-        if (!isSupported) {
-            return;
-        }
         //to be fixed
         return idb.init('db2-update', [
             {
@@ -108,9 +98,6 @@ describe('data-store/indexed-db/update', function() {
         });
     });
     it('doesn\'t let you update to a value where indexed fields are null', function(done) {
-        if (!isSupported) {
-            done();
-        }
         //to be fixed
         return idb.init('db3-update', [
             {
@@ -156,9 +143,6 @@ describe('data-store/indexed-db/update', function() {
         });
     });
     it('throws an error if an update is attempted on a non existing table', function(done) {
-        if (!isSupported) {
-            done();
-        }
         //to be fixed
         return idb.init('db4-update', [
             {
@@ -194,9 +178,6 @@ describe('data-store/indexed-db/update', function() {
         });
     });
     it('doesn\'t update anything if key search in index doesn\'t fetch any data', function() {
-        if (!isSupported) {
-            return;
-        }
         //to be fixed
         return idb.init('db5-update', [
             {
